test: :white_check_mark: Adding test case for tetsing the project

+++ import_doc.py
@@ -11,9 +11,9 @@ output_dir = root_path + "/data/pdf/"
 
 
 def main(
-    user_name,
-    job_title,
-    company_name,
+    user_name:str,
+    job_title:str,
+    company_name:str,
 ):
     if os.path.isdir(root_path + "/data/"):
         shutil.rmtree(root_path + "/data/")
@@ -21,41 +21,48 @@ def main(
         pass
     else:
         os.makedirs(output_dir)
-    template_file_path = folder_path + f"{job_title.lower()}_cover_letter.docx"
-    output_file_path = root_path + "/data/" + f"{user_name}-cover_letter.docx"
 
-    current_date = datetime.now().strftime("%d %B, %Y")
+    user_name=user_name.title()
+    files = [f for f in os.listdir(folder_path)]
+    available_job_title = [file.split("_")[0] for file in files]
+    if job_title.lower() in available_job_title:
+        template_file_path = folder_path + f"{job_title.lower()}_cover_letter.docx"
+        output_file_path = root_path + "/data/" + f"{user_name}-cover_letter.docx"
 
-    variables = {
-        "${DATE}": current_date,
-        "${JOB_TITLE}": job_title + "Developer",
-        "${COMPANY_NAME}": company_name,
-        "${YOUR_NAME}": user_name,
-    }
-    template_document = Document(template_file_path)
-    try:
-        for variable_key, variable_value in variables.items():
-            for paragraph in template_document.paragraphs:
-                replace_text_in_paragraph(paragraph, variable_key, variable_value)
+        current_date = datetime.now().strftime("%d %B, %Y")
 
-            for table in template_document.tables:
-                for col in table.columns:
-                    for cell in col.cells:
-                        for paragraph in cell.paragraphs:
-                            replace_text_in_paragraph(
-                                paragraph, variable_key, variable_value
-                            )
+        variables = {
+            "${DATE}": current_date,
+            "${JOB_TITLE}": job_title + "Developer",
+            "${COMPANY_NAME}": company_name,
+            "${YOUR_NAME}": user_name,
+        }
+        template_document = Document(template_file_path)
+        try:
+            for variable_key, variable_value in variables.items():
+                for paragraph in template_document.paragraphs:
+                    replace_text_in_paragraph(paragraph, variable_key, variable_value)
 
-        template_document.save(output_file_path)
-        doc2pdf(output_file_path)
-        if os.path.exists(output_file_path):
-            os.remove(output_file_path)
-            return {
-                "file_location": output_dir + f"{user_name}-cover_letter.pdf",
-                "file_name": f"{user_name}-cover_letter.pdf",
-            }
-    except Exception as e:
-        raise e
+                for table in template_document.tables:
+                    for col in table.columns:
+                        for cell in col.cells:
+                            for paragraph in cell.paragraphs:
+                                replace_text_in_paragraph(
+                                    paragraph, variable_key, variable_value
+                                )
+
+            template_document.save(output_file_path)
+            doc2pdf(output_file_path)
+            if os.path.exists(output_file_path):
+                os.remove(output_file_path)
+                return {
+                    "file_location": output_dir + f"{user_name}-cover_letter.pdf",
+                    "file_name": f"{user_name}-cover_letter.pdf",
+                }
+        except Exception as e:
+            raise e
+    else:
+        raise Exception(f"Job Title {job_title} not found.Available Job Title are {available_job_title}")
 
 
 def replace_text_in_paragraph(paragraph, key, value):
